Optimized string rendering with shadows. Optimized some strings rendering in gui and hud. Added ability to create string with shadows instead of realtime rendering. Refactoring

+++ engine/lwjgl3/src/main/java/net/bfsr/engine/renderer/font/StringParams.java
@@ -8,7 +8,7 @@ import org.joml.Vector4f;
 @Setter
 class StringParams {
     private Vector4f color = new Vector4f();
-    private int x, y;
+    private float x, y;
     private int height;
 
     void addHeight(int height) {
